+++ src/yt_maint_mercenaries.c
@@ -160,7 +160,8 @@ yt_maintenance_maintain_mercenary_base(struct yt_game *game,
 		struct yt_sector sector;
 		int today;
 
-		if (!yt_current_date_serial(&game->clock, game->config.epoch_year,
+		if (!yt_current_date_serial(&game->clock,
+		    (float)game->config.epoch_year,
 		    &today, NULL, error))
 			return false;
 		if (!yt_game_read_planet(game, planet_number, &planet, error))
